requirements.txt
Django==6.0
djangorestframework==3.16.1
coreapi==2.3.3
